fix: should disable find when editor is activated
